+++ changelog.md
@@ -2,3 +2,4 @@
 
 * Support update multiple gradle version catalogs.
 * If repositories not provided default set of repositories will be used (Maven Central, Google, Gradle Plugin Portal).
+* Fix: correct handle absence of `latest`, `release`, `lastUpdated` tags and just one version of dependency in artifact metadata.